refactor: move constants.hpp to Constants.hpp
- Move src/Editor/constants/constants.hpp to src/Constants/Constants.hpp
