+++ src/CMakeLists.txt
@@ -12,6 +12,7 @@ target_include_directories(kilo
 
 target_sources(kilo
     PRIVATE
+        "${PROJECT_SOURCE_DIR}/src/Constants/Constants.hpp"
         "${PROJECT_SOURCE_DIR}/src/Editor/cursor/cursor.hpp"
 
         "${PROJECT_SOURCE_DIR}/src/Editor/Editor.hpp"
